Change return type of init function and add call to import_array
Any C extension that uses the Numpy C library must call the
import_array function in its init function or else it will not work
correctly. I left this out when modifying the init functio, so this
commit restores it.

I also changed the return type of the init function to use the
PyMODINIT_FUNC macro. While not strictly necessary, it is a
recommended proactice in the Numpy documentation.