Correctly handle pixmaps containing NaN values
Some instruments do not define a coordinate transformation across the
entire image. In these cases the pixmap which is passed to drizzle
contains NaN values in some pixels. Usually these pixels are around
the edge of the image. Since the drizzle code was not written to
handle the case of NaN values, it needed to be modified.

Two modifications were made. Now the code only drizzles image areas
where the pixmap contains defined values. The change treats areas in
the input image where pixmap contains NaNs as if they do not overlap
the output image. This change is in a new function, shrink_segment,
which is called by clip_bounds.

Also, the bilinear interpolation code which handle the coordinate
transformation at fractional pixel values excludes pixels with NaN
values. This typically will happen at the ragged boundary between
defined and undefined values. This change is in a rewritten version of
map_point, which calls a new function, map_bounds. This function does
a single spiral search to find the four points nearest the point to be
interpolated at.