Add check to input of shrinks_segment to prevent checks outside of array
The function shrink_segment in cdrizzlemap.c reduces the segment
length to exclude NaN values on either end of the segment. The old version
did not check to see if the segment lies entirely inside the bounds of the
image, and sometimes it is called with segments like this. I added checks
so the the search for NaN vaules starts on the image and not outside it.