Fix out of bounds array error caused by shrink_segment
Drizzle computes the overlap between each line of the input image and
the output image. Part of the computation is shrink_segment, which
reduces the length of the input image line if there are NaN values
along the edge of the image. That code mistakenly assumed that the
size of the input image and pixel map were the same, leading to the
out of bounds error when the pixel map was larger. The code in
shrink_segment has been modified to intialize the line length from the
input image and not the pixel map.

Some logging code has been added to cdrizzleutil.h and various
functions to make catching errors easier in the future. The code is
controlled by a C macro, LOGGING, and is usually turned off. There is
a new global variable, driz_log_handle, defined in cdrizzleapi,
initiallized to NULL.