Fix logic of shrink_segment
Before drizzling input pixels onto the output image, drizzle must
first determine what part of each input line overlaps the output
image. And since lines typically have invalid values at their edges,
we would rather deal with the valid portion of the input line than the
whole line. Shrink_segment walks inward from the edges of the line to
find the first valid value, that is, a value that is not NaN. The
logic of the code was previously badly wrong, leading to a
segmentation error when the size of the input and output images
differed. The revised code fixes the logic error and changes the
parmeter list of shrink_segment.