Add pixmap dimension check to drizzle
The drizzle code expects that the dimensions of the input image and
pixmap file are the same. The blotting code expects the dimensions of
the output image and pixmap file are the same. Prior to this change,
the C code would throw a segmentation violation if the pixmap did not
meet this condition. This change adds a check and throws an explicit
error if the condition is not met. If your pixmap is larger than the
input/output image, it can be trimmmed to match the dimensions. If it
is smaller, resize it to have the correct dimensions and fill the
undefined pixels with NaN and the function shrink_segment will see
that only the corresponding section of the input image is used in the
drizzle. Some diagnostic prints used in debugging C errors are also
modified in this change. They are turned off in production. This new
test may explicitly break existing code, but it was going to fail
anyway with a harder to debug segmentation violation.