Modify shrink_segment to check weights array for zero values
The purpose of shrink_segment is to extract the segment of the input
line where the input data is valid. Previously it checked to see if
pixmap had a vaild value. Now it also checks if the weights array is
non_zero, as zero values are used to flag invalid data. The intention
is to speed up the drizzle code by omitting calculations that will
not contribute to the result.