Check for weights array and only shrink segment for lines
The weights array is optional. If null, all weights are assumed to be
one. Before checking it or its dimensions the code needs to check if
the weights array is null.

The function shrink_segment is run when computing the overlap between
the input image line and output image. It's incorrect to apply it when
computing the image overlap. The code has been re-arranged so this
doesn't happen.