Have proper bounds when interpolating in pixmap or fail gracefully
A recent change in the pipeline code has resulted in the pixmap only
being partially populated with values mapping from the input to output
images. The remainder of the pixmap is populated with NaNs. When
mapping a point between the input and output image, one must first
find a bound, a set of four pixels that allows one to do an
interpolation or extrapolation in two dimensions. The bound does not
need to be the immediate neighbors of the interpolated point, but does
need to be in the bordering rows and columns of the point. For this
reason the interpolation fails if a bordering row or column is
entirely NaNs. This typically happens when spectra are being drizzled,
the defined area of the pixmap is only a subregion of the detector
image.

To handle this common situation, the drizzle code was modified by
rewriting shrink_segment to compute the rectangle holding all the
defined values in the pixmap. This rectangle can also hold some
undefined values (NaNs), so the code to calculate the interpolation
bounds (now named interpolation_bounds) was rewritten to recognize the
case when the search for a bounds fails and return a flag indicating
the failure. (One indicates failure, zero indicates success.) This
flag "bubbles up" to do_kernel_square, where it is checked for, and if
found, the pixel is skipped and marked as a miss in the drizzle
calculation. The other kernels in cdrizzlebox were also modified to
skip flagged pixels.

In addition the logging code in cdrizzleutil.h was modified to make it
easier to debug this change. And some code changes were made in order
to allow the flag to bubble up correctly.