Support NaN as a fill value
The drizzle user has the option of setting unwritten pixels in the
output image to a fill value rather than the default value of
zero. This value is set by the drizzle parameter fillstr. If this
parameter is "NaN" or "nan" the fill value in the output image will be
"Not a Number."